+++ Celebrity/app/helpers/information_helper.rb
@@ -1,0 +1,2 @@
+module InformationHelper
+end
