+++ Celebrity/config/routes.rb
@@ -54,7 +54,9 @@ Rails.application.routes.draw do
   delete '/delete_tag', to: 'users#tag_delete' #temp post⇨deleteに変更しshow.htmlと統一
   patch '/post_skillsheet', to: 'skillsheets#update_skillsheet'
   root 'top#index'
-  post 'top#index', to: 'top#update_information', as: 'update_information'
+  
+  get '/information', to: 'information#show'
+  post '/information', to: 'information#create', as: 'create_information'
 
   get '/progate', to: 'static_pages#progate'
   post '/progate', to: 'static_pages#progate'
